% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_estimator.R
\name{compute_corr}
\alias{compute_corr}
\title{Computes a tidy correlation}
\usage{
compute_corr(data, var1, var2)
}
\arguments{
\item{input}{data set}

\item{name}{of variable 2}
}
\value{
A tibble with the Pearson correlation and the p-value
}
\description{
Computes a tidy correlation
}
\examples{
compute_corr(data = faithful, var1 = eruptions, var2 = waiting)
}
