% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Transform helper}
\usage{
foo(df)
}
\arguments{
\item{df}{A data frame with columns x and y.}
}
\value{
A data frame with columns x, y, and z (where z = x + y).
}
\description{
Example using dplyr without pkg:: qualification.
}
\examples{
df <- data.frame(x = 1:5, y = 6:10, a = 11:15)
foo(df)
}
