% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goo.R
\name{goo}
\alias{goo}
\title{Goo Function
This function performs a specific operation on the input `x` using an internal helper function.}
\usage{
goo(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector after applying the internal function.
}
\description{
Goo Function
This function performs a specific operation on the input `x` using an internal helper function.
}
